requirements.txt
pandas[excel]
streamlit
streamlit-authenticator
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
Authlib